fix(288): the close action that two of four SLTP venues ignored
`include_close_action` puts ("close", None, None) in the action map for all four
futures SLTP venues -- the field is on the shared `BaseFuturesSLTPConfig`. Only
bybit and okx ever handled it. On binance and bitget the tuple fell past both
`if side == "long"` and `elif side == "short"` and returned executed=False: the
policy asked to close, the position stayed open, and nothing said so.

    venue     close_position called?   position after
    binance   False   executed=False   1   <-- close ignored
    bitget    False   executed=False   1   <-- close ignored
    bybit     True    executed=True    0
    okx       True    executed=True    0

It defaults False, so it bites only a user who enables it -- but it is enabled by
a config field, not by editing code, and the failure is silent.

The fold is what made it visible. Replacing binance's hand-rolled bracket-price
arithmetic with `calculate_bracket_prices` -- which bitget already called -- left
the two `_execute_trade_if_needed` bodies 100% identical on code lines (149 vs
152 lines, differing only by three of bitget's comments). One copy now lives on
SLTPMixin. bybit and okx keep their overrides: the four venues split by SIGNATURE,
not by exchange, since those two take the mark threaded in by the caller
(`*, current_price`, #295) rather than pricing off their own candle.

Carrying bybit's guard across needed its early return, not just its side
derivation: `"buy" if side == "long" else "sell"` alone turns the silent no-op
into a wrong SELL on an open short. The regression test is parametrised over both
directions for that reason, and asserts `close_position` actually reached the
exchange -- a trade_info flag alone passes on an env that reports success without
trading.

Net -84 lines. Suite 4405 passed.
