+++ torchtrade/envs/utils/sltp_mixin.py
@@ -1,11 +1,14 @@
 """Shared SLTP (Stop-Loss/Take-Profit) functionality for live trading environments."""
 
 import logging
+import math
+from typing import Dict, Optional, Tuple
 
 import torch
 from tensordict import TensorDictBase
 
 from torchtrade.envs.core.state import position_direction_from_status
+from torchtrade.envs.utils.sltp_helpers import calculate_bracket_prices
 
 logger = logging.getLogger(__name__)
 
@@ -167,3 +170,184 @@ class SLTPMixin:
         """
         self.active_stop_loss = 0.0
         self.active_take_profit = 0.0
+
+    def _execute_trade_if_needed(
+        self, action_tuple: Tuple[Optional[str], Optional[float], Optional[float]]
+    ) -> Dict:
+        """Place the bracket. Two byte-identical copies (binance, bitget) folded here (#288).
+
+        The four SLTP venues split by SIGNATURE, not by exchange: binance and bitget price
+        the bracket off the observer's own candle, while bybit and okx require the mark
+        threaded in by the caller (`*, current_price`) -- see #295, where re-reading it
+        inside the trade path bypassed the halt policy. Those two override this.
+
+        Args:
+            action_tuple: (side, stop_loss_pct, take_profit_pct); (None, None, None) is HOLD.
+
+        Returns:
+            Dict with trade execution info
+        """
+        trade_info = {
+            "executed": False,
+            "quantity": 0,
+            "side": None,
+            "success": None,
+            "closed_position": False,
+        }
+
+        side, stop_loss_pct, take_profit_pct = action_tuple
+
+        # HOLD action - do nothing
+        if side is None:
+            return trade_info
+
+        # Position locking: ignore all actions while in position
+        if self.config.lock_position_until_sltp and self.position.current_position != 0:
+            return trade_info
+
+        # `include_close_action` puts ("close", None, None) in the action map for all four
+        # venues (sltp_config.py), but only bybit/okx ever handled it: on binance/bitget it
+        # fell past both `side == "long"` and `elif side == "short"` and returned
+        # executed=False. The policy asked to close, the position stayed open, and nothing
+        # said so. Sharing one copy is what made that visible (#288).
+        if side == "close":
+            if self.position.current_position == 0:
+                return trade_info
+            try:
+                success = self.trader.close_position()
+            except Exception as e:
+                logger.error(f"Close position failed for {self.config.symbol}: {e}")
+                return trade_info
+            if success:
+                # A realised close moves equity; the cached balance is now wrong by the
+                # trade's P&L. SUCCESS only -- a failed close leaves the position (#295).
+                self._last_confirmed_read.pop("balance", None)
+                close_side = "sell" if self.position.current_position > 0 else "buy"
+                self.position.current_position = 0
+                self.active_stop_loss = 0.0
+                self.active_take_profit = 0.0
+                trade_info.update({
+                    "executed": True, "side": close_side,
+                    "success": True, "closed_position": True,
+                })
+            return trade_info
+
+        # Check if already in same position (ignore duplicate actions)
+        if side in self.SIDE_DIRECTION and self.position.current_position == self.SIDE_DIRECTION[side]:
+            return trade_info  # Already in this position, ignore duplicate action
+
+        # Under `_halting`, read AND verdict (#295). `get_observations` raises ValueError
+        # on a short window or a stale last bar, and this sat outside the policy -- so a
+        # degraded feed escaped as a bare ValueError in EVERY trade_mode, since this runs
+        # before the mode branch. bybit/okx take a threaded mark instead; these two price
+        # brackets off the candle close deliberately, so the read stays rather than being
+        # replaced by the mark.
+        def read_close():
+            obs = self.observer.get_observations(return_base_ohlc=True)
+            current_price = float(obs["base_features"][-1, 3])
+            # This price divides the notional sizing AND prices both brackets in every
+            # mode, including the "quantity" default which checked nothing. dropna() does
+            # not clear a candle close of inf (#347). The name is load-bearing:
+            # test_sltp_sizing_rejects_a_non_finite_price_or_balance greps for it.
+            if not math.isfinite(current_price) or current_price <= 0:
+                raise ValueError(
+                    f"unusable close price ({current_price}) for {self.config.symbol}"
+                )
+            return current_price
+
+        # cache_key is load-bearing, not decoration: without it `cached` is None, grace
+        # cannot apply, and this still raises -- it just raises a nicer type. The claimed
+        # behaviour is "serve the last CONFIRMED close and flag the bar", which needs a
+        # slot to serve from. Its own slot, because it is a candle close, not the mark.
+        current_price = self._halting(read_close, cache_key="candle_close")
+
+        quantity = self._resolve_bracket_quantity(current_price)
+        if quantity is None:
+            trade_info["success"] = False
+            return trade_info
+
+        # Close opposite position if switching directions
+        if self.position.current_position != 0:
+            # We have an existing position that needs to be closed before opening new one
+            if (side == "long" and self.position.current_position == -1) or \
+               (side == "short" and self.position.current_position == 1):
+                try:
+                    close_success = self.trader.close_position()
+                except Exception as e:
+                    logger.error(f"Close position failed for {self.config.symbol}: {e}")
+                    return trade_info
+                if not close_success:
+                    return trade_info
+                # A realised close moves equity; the cached balance is now wrong by the
+                # trade's P&L. Reached only on success (#295).
+                self._last_confirmed_read.pop("balance", None)
+                self.position.current_position = 0
+
+        if side == "long":
+            # Open LONG with SL/TP bracket order
+            stop_loss_price, take_profit_price = calculate_bracket_prices(
+                "long", current_price, stop_loss_pct, take_profit_pct
+            )
+
+            try:
+                success = self.trader.trade(
+                    side="buy",
+                    quantity=quantity,
+                    order_type="market",
+                    take_profit=take_profit_price,
+                    stop_loss=stop_loss_price,
+                )
+
+                if success:
+                    # Only record SL/TP levels that actually placed on-exchange
+                    bs = getattr(self.trader, 'bracket_status', {"tp_placed": True, "sl_placed": True})
+                    self.active_stop_loss = stop_loss_price if bs["sl_placed"] else 0.0
+                    self.active_take_profit = take_profit_price if bs["tp_placed"] else 0.0
+
+                trade_info.update({
+                    "executed": True,
+                    "quantity": quantity,
+                    "side": "buy",
+                    "success": success,
+                    "stop_loss": stop_loss_price,
+                    "take_profit": take_profit_price,
+                })
+            except Exception as e:
+                logger.error(f"Long trade failed for {self.config.symbol}: quantity={quantity}, SL={stop_loss_price:.2f}, TP={take_profit_price:.2f}, error={e}")
+                trade_info["success"] = False
+                return trade_info
+
+        elif side == "short":
+            # Open SHORT with SL/TP bracket order
+            stop_loss_price, take_profit_price = calculate_bracket_prices(
+                "short", current_price, stop_loss_pct, take_profit_pct
+            )
+
+            try:
+                success = self.trader.trade(
+                    side="sell",
+                    quantity=quantity,
+                    order_type="market",
+                    take_profit=take_profit_price,
+                    stop_loss=stop_loss_price,
+                )
+
+                if success:
+                    bs = getattr(self.trader, 'bracket_status', {"tp_placed": True, "sl_placed": True})
+                    self.active_stop_loss = stop_loss_price if bs["sl_placed"] else 0.0
+                    self.active_take_profit = take_profit_price if bs["tp_placed"] else 0.0
+
+                trade_info.update({
+                    "executed": True,
+                    "quantity": quantity,
+                    "side": "sell",
+                    "success": success,
+                    "stop_loss": stop_loss_price,
+                    "take_profit": take_profit_price,
+                })
+            except Exception as e:
+                logger.error(f"Short trade failed for {self.config.symbol}: quantity={quantity}, SL={stop_loss_price:.2f}, TP={take_profit_price:.2f}, error={e}")
+                trade_info["success"] = False
+                return trade_info
+
+        return trade_info
